backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
python-jose==3.3.0
cryptography==41.0.7
